backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
numpy>=1.20.0
scipy>=1.7.0
pyyaml>=5.4.0
